pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "asyncio>=3.4.3",
    "requests>=2.32.3",
    "streamlit>=1.43.2",
    "trafilatura>=2.0.0",
    "crawl4ai>=0.1.1",
    "pyppeteer>=1.0.2",
    "openai>=1.1.0",
    "fastapi>=0.109.1",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0"
]